Fixed compile error caused by duplicate class names
org.modelix.model.PlatformSpecific was specified in model-api
and model-client.
