+++ model-api/src/commonMain/kotlin/org/modelix/model/api/COWArrays.kt
@@ -3,6 +3,19 @@
  * you may not use this file except in compliance with the License.
  * You may obtain a copy of the License at
  *
+ * http://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS,
+ * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ * See the License for the specific language governing permissions and
+ * limitations under the License.
+ */
+/*
+ * Licensed under the Apache License, Version 2.0 (the "License");
+ * you may not use this file except in compliance with the License.
+ * You may obtain a copy of the License at
+ *
  *  http://www.apache.org/licenses/LICENSE-2.0
  *
  * Unless required by applicable law or agreed to in writing,
@@ -13,7 +26,7 @@
  * under the License. 
  */
 
-package org.modelix.model.util.pmap
+package org.modelix.model.api
 
 object COWArrays {
