+++ model-api/src/commonMain/kotlin/org/modelix/model/api/ContextNodeResolveContext.kt
@@ -15,7 +15,7 @@
 
 package org.modelix.model.api
 
-import org.modelix.model.util.ContextValue
+import org.modelix.model.api.ContextValue
 
 object ContextNodeResolveContext {
     val CONTEXT_VALUE = ContextValue<INodeResolveContext>()
