+++ model-api/src/commonMain/kotlin/org/modelix/model/api/PlatformSpecific.kt
@@ -1,4 +1,4 @@
-package org.modelix.model
+package org.modelix.model.api
 
 import kotlin.reflect.KClass
