+++ model-api/src/commonMain/kotlin/org/modelix/model/area/NodeChangeEvent.kt
@@ -1,6 +1,4 @@
 /*
- * Copyright 2003-2020 JetBrains s.r.o.
- *
  * Licensed under the Apache License, Version 2.0 (the "License");
  * you may not use this file except in compliance with the License.
  * You may obtain a copy of the License at
