--- model-api/src/commonMain/kotlin/org/modelix/model/util/ContextValue.kt
@@ -1,11 +1,0 @@
-package org.modelix.model.util
-
-expect class ContextValue<E> {
-
-    constructor()
-    constructor(defaultValue: E)
-
-    fun getValue(): E?
-    fun runWith(newValue: E, r: () -> Unit)
-    fun <T> computeWith(newValue: E, r: () -> T): T
-}
