+++ model-api/src/jsMain/kotlin/org/modelix/model/api/ContextValue.kt
@@ -1,0 +1,44 @@
+/*
+ * Licensed under the Apache License, Version 2.0 (the "License");
+ * you may not use this file except in compliance with the License.
+ * You may obtain a copy of the License at
+ *
+ * http://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS,
+ * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ * See the License for the specific language governing permissions and
+ * limitations under the License.
+ */
+package org.modelix.model.api
+
+actual class ContextValue<E> {
+    private var value: E?
+
+    actual constructor() {
+        value = null
+    }
+
+    actual constructor(defaultValue: E) {
+        value = defaultValue
+    }
+
+    actual fun getValue(): E? {
+        return value
+    }
+
+    actual fun runWith(newValue: E, r: () -> Unit) {
+        computeWith(newValue, r)
+    }
+
+    actual fun <T> computeWith(newValue: E, r: () -> T): T {
+        val oldValue = value
+        value = newValue
+        try {
+            return r()
+        } finally {
+            value = oldValue
+        }
+    }
+}
