+++ model-api/src/jvmMain/kotlin/org/modelix/model/api/PlatformSpecific.kt
@@ -1,4 +1,17 @@
-package org.modelix.model
+/*
+ * Licensed under the Apache License, Version 2.0 (the "License");
+ * you may not use this file except in compliance with the License.
+ * You may obtain a copy of the License at
+ *
+ * http://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS,
+ * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ * See the License for the specific language governing permissions and
+ * limitations under the License.
+ */
+package org.modelix.model.api
 
 import org.apache.commons.collections4.map.LRUMap
 import org.apache.log4j.Logger
