+++ model-client/src/commonMain/kotlin/org/modelix/model/client/GarbageFilteringStore.kt
@@ -18,7 +18,7 @@ package org.modelix.model.client
 import org.modelix.model.IKeyListener
 import org.modelix.model.IKeyValueStore
 import org.modelix.model.persistent.HashUtil
-import org.modelix.model.runSynchronized
+import org.modelix.model.api.runSynchronized
 
 class GarbageFilteringStore(private val store: IKeyValueStore) : IKeyValueStore {
     private val pendingEntries: MutableMap<String?, String?> = HashMap()
