+++ model-client/src/commonMain/kotlin/org/modelix/model/operations/OTBranch.kt
@@ -17,7 +17,7 @@ package org.modelix.model.operations
 
 import org.modelix.model.api.*
 import org.modelix.model.lazy.IDeserializingKeyValueStore
-import org.modelix.model.runSynchronized
+import org.modelix.model.api.runSynchronized
 
 class OTBranch(
     private val branch: IBranch,
