+++ model-client/src/commonMain/kotlin/org/modelix/model/operations/OTWriteTransaction.kt
@@ -21,10 +21,9 @@ import org.modelix.model.api.IIdGenerator
 import org.modelix.model.api.INodeReference
 import org.modelix.model.api.ITree
 import org.modelix.model.api.IWriteTransaction
-import org.modelix.model.api.PNodeAdapter
 import org.modelix.model.lazy.DuplicateNodeId
 import org.modelix.model.lazy.IDeserializingKeyValueStore
-import org.modelix.model.logTrace
+import org.modelix.model.api.logTrace
 
 class OTWriteTransaction(
     private val transaction: IWriteTransaction,
