+++ model-client/src/jvmMain/kotlin/org/modelix/model/util/pmap/LongKeyPMap.kt
@@ -16,9 +16,9 @@
 package org.modelix.model.util.pmap
 
 import org.apache.commons.lang3.mutable.MutableObject
-import org.modelix.model.util.pmap.COWArrays.insert
-import org.modelix.model.util.pmap.COWArrays.removeAt
-import org.modelix.model.util.pmap.COWArrays.set
+import org.modelix.model.api.COWArrays.insert
+import org.modelix.model.api.COWArrays.removeAt
+import org.modelix.model.api.COWArrays.set
 import java.util.function.BiPredicate
 
 class LongKeyPMap<V> protected constructor(root: INode<V?>?) {
